+++ de.metas.adempiere.adempiere/base/src/main/java-legacy/de/metas/impexp/excel/ArrayExcelExporter.java
@@ -11,7 +11,7 @@
  * with this program; if not, write to the Free Software Foundation, Inc., *
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA. *
  *****************************************************************************/
-package org.adempiere.impexp;
+package de.metas.impexp.excel;
 
 import java.util.List;
 import java.util.Properties;
