+++ de.metas.adempiere.adempiere/base/src/main/java-legacy/de/metas/impexp/excel/CellValueType.java
@@ -1,4 +1,4 @@
-package org.adempiere.impexp;
+package de.metas.impexp.excel;
 
 /*
  * #%L
